Added QuizQuestion Entity

+++ src/Nabeey.Domain/Entities/Articles/Article.cs
@@ -6,6 +6,6 @@ namespace Nabeey.Domain.Entities.Articles;
 public class Article : Auditable
 {
     public string Text { get; set; }
-    public int ContentId { get; set; }
+    public long ContentId { get; set; }
     public Content Content { get; set; }
 }
